requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
opencv-python>=4.5.3
Pillow>=8.3.2
fastapi>=0.68.1
uvicorn>=0.15.0
python-multipart>=0.0.5
requests>=2.26.0
scikit-image>=0.18.3
albumentations>=1.1.0
